pyproject.toml
[project]
name = "qlunee-code"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "anthropic>=0.95.0",
    "pytest>=9.0.3",
    "python-dotenv>=1.2.2",
    "pyyaml>=6.0.3",
    "reportlab>=4.4.10",
]
